Refactor the countdown timer
